==> Head/2Scripts/01_MtDnaFullList.ipynb <==
import pandas as pd
import Bio as bio
from Bio import SeqIO
from Bio import GenBank
import re
import numpy as np

print (bio.__version__)

open('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/1Raw/mitochondrion.genomic.gbff',"w").write(open('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/1Raw/mitochondrion.1.genomic.gbff',"r").read() + open('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/1Raw/mitochondrion.2.genomic.gbff',"r").read())

ids = [seq_record.id for seq_record in SeqIO.parse('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/1Raw/mitochondrion.genomic.gbff', "genbank")]

print(len(ids))

d = {"RefSeqID":pd.Series(ids, index=range(0,len(ids)))}
MtDnaFullList = pd.DataFrame(d)

MtDnaFullList

MtDnaFullList['SpeciesNames'] = [seq_record.annotations["organism"] for seq_record in SeqIO.parse('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/1Raw/mitochondrion.genomic.gbff', "genbank")]

MtDnaFullList['Taxonomy'] = [seq_record.annotations["taxonomy"] for seq_record in SeqIO.parse('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/1Raw/mitochondrion.genomic.gbff', "genbank")]

MtDnaFullList.head()

MtDnaFullList.to_csv('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/2Derived/01_MtDnaFullList.csv', sep=";")

## подсчет геномных характеристик

MtDnaFullList['Lenght'] = [len(seq_record) for seq_record in SeqIO.parse('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/1Raw/mitochondrion.genomic.gbff', "genbank")]

MtDnaFullList['Num_A'] = [seq_record.seq.count("A") for seq_record in SeqIO.parse(r"C:\Users\PC\Desktop\MitoClub\mtDnaInvertebrate\Body\1Raw\mitochondrion.genomic.gbff", "genbank")]
MtDnaFullList['Num_T'] = [seq_record.seq.count("T") for seq_record in SeqIO.parse(r"C:\Users\PC\Desktop\MitoClub\mtDnaInvertebrate\Body\1Raw\mitochondrion.genomic.gbff", "genbank")]
MtDnaFullList['Num_G'] = [seq_record.seq.count("G") for seq_record in SeqIO.parse(r"C:\Users\PC\Desktop\MitoClub\mtDnaInvertebrate\Body\1Raw\mitochondrion.genomic.gbff", "genbank")]
MtDnaFullList['Num_C'] = [seq_record.seq.count("C") for seq_record in SeqIO.parse(r"C:\Users\PC\Desktop\MitoClub\mtDnaInvertebrate\Body\1Raw\mitochondrion.genomic.gbff", "genbank")]

MtDnaFullList['%A'] = 100 * MtDnaFullList['Num_A'] / MtDnaFullList['Lenght']
MtDnaFullList['%T'] = 100 * MtDnaFullList['Num_T'] / MtDnaFullList['Lenght']
MtDnaFullList['%G'] = 100 * MtDnaFullList['Num_G'] / MtDnaFullList['Lenght']
MtDnaFullList['%C'] = 100 * MtDnaFullList['Num_C'] / MtDnaFullList['Lenght']

MtDnaFullList['GC'] = MtDnaFullList['%G'] + MtDnaFullList['%C']

MtDnaFullList = MtDnaFullList.drop(np.where(MtDnaFullList['GC'] == 0)[0])

MtDnaFullList['log2Lenght'] = np.log2(MtDnaFullList['Lenght'])

## отдельный файл с таксономией

MtDnaFullList = pd.read_csv(r'C:\Users\PC\Desktop\MitoClub\MtDnaInvertebrate\Body\2Derived\MtDnaFullList.csv', sep="\t")

new_df = MtDnaFullList['Taxonomy'].str.split(', ',expand=True)

new_df.head()

new_df[1].value_counts()

new_df[2].value_counts()

new_df.to_csv(r'C:\Users\PC\Desktop\MitoClub\MtDnaInvertebrate\Body\2Derived\Taxonomy.csv', sep="\t")

human = SeqIO.parse('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/1Raw/sequence.gb', format='genbank')
for seq in human:
    for feature in seq.features:
        if feature.type == 'CDS' and feature.strand != 1:
            print(feature)
            print(feature.extract(seq.seq))
