==> Head/2Scripts/03_AddGenomeParameters.ipynb <==
import pandas as pd
import Bio as bio
from Bio import SeqIO
from Bio import GenBank
import re
import numpy as np

Tax = pd.read_csv('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/2Derived/02_DeriveOfTaxonomyGroups.csv', sep=";")

Tax.head()

sequences = SeqIO.parse('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/1Raw/mitochondrion.genomic.gbff', "genbank")
for seq_record in sequences:
    try:
        idx = seq_record.id
        Tax.loc[Tax['RefSeqID'] == idx, 'Length'] = len(seq_record.seq)
        Tax.loc[Tax['RefSeqID'] == idx, 'Num_A'] = seq_record.seq.count('A')
        Tax.loc[Tax['RefSeqID'] == idx, 'Num_T'] = seq_record.seq.count('T')
        Tax.loc[Tax['RefSeqID'] == idx, 'Num_G'] = seq_record.seq.count('G')
        Tax.loc[Tax['RefSeqID'] == idx, 'Num_C'] = seq_record.seq.count('C')
    except:
        idx = seq_record.id
        Tax.loc[Tax['RefSeqID'] == idx, 'Length'] = 'NaN'
        Tax.loc[Tax['RefSeqID'] == idx, 'Num_A'] = 'NaN'
        Tax.loc[Tax['RefSeqID'] == idx, 'Num_T'] = 'NaN'
        Tax.loc[Tax['RefSeqID'] == idx, 'Num_G'] = 'NaN'
        Tax.loc[Tax['RefSeqID'] == idx, 'Num_C'] = 'NaN'

Tax.drop(['X'], axis='columns', inplace=True)

Tax = Tax.drop(np.where(Tax['Length'] == 'NaN')[0])

Tax.reset_index(drop=True, inplace=True)

Tax[['Length', 'Num_A', 'Num_T', 'Num_C', 'Num_G']] = Tax[['Length', 'Num_A', 'Num_T', 'Num_C', 'Num_G']].astype('int')

Tax

Tax['PrA'] = 100 * Tax['Num_A'] / Tax['Length']
Tax['PrT'] = 100 * Tax['Num_T'] / Tax['Length']
Tax['PrG'] = 100 * Tax['Num_G'] / Tax['Length']
Tax['PrC'] = 100 * Tax['Num_C'] / Tax['Length']

Tax['GC'] = Tax['PrG'] + Tax['PrC']

Tax.to_csv('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/2Derived/03_AddGenomeParameters.csv', sep=";")

Tax = pd.read_csv('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/2Derived/03_AddGenomeParameters.csv', sep=";")

Tax

Tax['Group3'] = 'Eukaryota'
Tax['Group3'].loc[Tax['Taxonomy'].str.contains('Fungi', na=False)] = 'Fungi'
Tax['Group3'].loc[Tax['Taxonomy'].str.contains('Viridiplantae', na=False)] = 'Plants'
Tax['Group3'].loc[Tax['Taxonomy'].str.contains('Metazoa', na=False)] = 'Invertebrates'
Tax['Group3'].loc[Tax['Taxonomy'].str.contains('Vertebrata', na=False)] = 'Vertebrata'
Tax['Group3'].loc[Tax['Taxonomy'].str.contains('Vertebrata' and 'Chondrichthyes', na=False)] = 'Chondrychthyes'
Tax['Group3'].loc[Tax['Taxonomy'].str.contains('Vertebrata' and 'Actinopterygii', na=False)] = 'Actinopterygii'
Tax['Group3'].loc[Tax['Taxonomy'].str.contains('Vertebrata' and 'Amphibia', na=False)] = 'Amphibia'
Tax['Group3'].loc[Tax['Taxonomy'].str.contains('Vertebrata' and 'Archelosauria')] = 'Reptilia'
Tax['Group3'].loc[Tax['Taxonomy'].str.contains('Vertebrata' and 'Lepidosauria', na=False)] = 'Reptilia'
Tax['Group3'].loc[Tax['Taxonomy'].str.contains('Vertebrata' and 'Aves', na=False)] = 'Aves'
Tax['Group3'].loc[Tax['Taxonomy'].str.contains('Vertebrata' and 'Mammalia', na=False)] = 'Mammalia'

Tax['GC'] = Tax['PrC'] + Tax['PrG']

Tax.groupby(['Group2'])['GC'].mean()
