==> Head/2Scripts/08_CodonUsageDerived.ipynb <==
import pandas as pd
import os
import tqdm

## Simple codon usage

result = pd.DataFrame(columns = ['Unnamed: 0', 'GenName', 'ProteinID', 'OrganismName', 'Taxonomy', 'Strand', 'TTT','TTC','TTA','TTG','TCT','TCC','TCA','TCG','TAT','TAC','TGT','TGC','TGA','TGG','CTT','CTC','CTA','CTG','CCT','CCC','CCA','CCG','CAT','CAC','CAA','CAG','CGT','CGC','CGA','CGG','ATT','ATC','ATA','ATG','ACT','ACC','ACA','ACG','AAT','AAC','AAA','AAG','AGT','AGC','AGA','AGG','GTT','GTC','GTA','GTG','GCT','GCC','GCA','GCG','GAT','GAC','GAA','GAG','GGT','GGC','GGA','GGG','TAA','TAG'])
result

direct = '/home/emulciber/MitoClub/mtDnaInvertebrate/Body/2Derived/codon_usage'
dirname = os.listdir(direct)

for file in tqdm.tqdm(dirname):
    try:
        df = pd.read_csv((os.path.join(direct, file)), sep='\t')
        result = result.append(df, ignore_index=True)
    except:
        pass

result

aa_mm = pd.read_csv('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/2Derived/AminoAcids_mismatched.csv', sep=';')
aa_mm

exclude_list = list(aa_mm['ProteinID'])

result_wt_mm = result.loc[~result['ProteinID'].isin(exclude_list)]

Tax = pd.read_csv('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/2Derived/02_DeriveOfTaxonomyGroups.csv', sep=';')

Tax

Tax = Tax.loc[:, ['SpeciesNames', 'Group1', 'Group2']]
Tax = Tax.rename(columns={'SpeciesNames': 'OrganismName'})

result_wt_mm = result_wt_mm.merge(Tax, how='left', on='OrganismName')

result_wt_mm['Group3'] = 'Eukaryota'
result_wt_mm['Group3'].loc[result_wt_mm['Taxonomy'].str.contains('Fungi', na=False)] = 'Fungi'
result_wt_mm['Group3'].loc[result_wt_mm['Taxonomy'].str.contains('Viridiplantae', na=False)] = 'Plants'
result_wt_mm['Group3'].loc[result_wt_mm['Taxonomy'].str.contains('Metazoa', na=False)] = 'Invertebrates'
result_wt_mm['Group3'].loc[result_wt_mm['Taxonomy'].str.contains('Vertebrata', na=False)] = 'Vertebrata'
result_wt_mm['Group3'].loc[result_wt_mm['Taxonomy'].str.contains('Vertebrata' and 'Chondrichthyes', na=False)] = 'Chondrychthyes'
result_wt_mm['Group3'].loc[result_wt_mm['Taxonomy'].str.contains('Vertebrata' and 'Actinopterygii', na=False)] = 'Actinopterygii'
result_wt_mm['Group3'].loc[result_wt_mm['Taxonomy'].str.contains('Vertebrata' and 'Amphibia', na=False)] = 'Amphibia'
result_wt_mm['Group3'].loc[result_wt_mm['Taxonomy'].str.contains('Vertebrata' and 'Archelosauria')] = 'Reptilia'
result_wt_mm['Group3'].loc[result_wt_mm['Taxonomy'].str.contains('Vertebrata' and 'Lepidosauria', na=False)] = 'Reptilia'
result_wt_mm['Group3'].loc[result_wt_mm['Taxonomy'].str.contains('Vertebrata' and 'Aves', na=False)] = 'Aves'
result_wt_mm['Group3'].loc[result_wt_mm['Taxonomy'].str.contains('Vertebrata' and 'Mammalia', na=False)] = 'Mammalia'

result_wt_mm

gainers = ['CCT', 'CCC', 'CCA', 'CCG', 'ACT', 'ACC', 'ACA', 'ACG', 'CAT', 'CAC', 'CAA', 'CAG', 'AAT', 'AAC', 'AAA', 'AAG']
losers = ['TTT', 'TTC', 'TTA', 'TTG', 'GTC', 'GTA', 'GTG', 'GTT', 'TGT', 'TGC', 'TGA', 'TGG', 'GGT', 'GGC', 'GGA', 'GGG']
intermediates = ['TCT', 'TCC', 'TCA', 'TCG', 'TAT', 'TAC', 'TAA', 'TAG', 'GCT', 'GCC', 'GCA', 'GCG', 'GAT', 'GAC', 'GAA', 'GAG', 'CTT', 'CTC', 'CTA', 'CTG', 'ATT', 'ATC', 'ATA', 'ATG', 'CGT', 'CGC', 'CGA', 'CGG', 'AGT', 'AGC', 'AGA', 'AGG']

result_wt_mm.loc[:, 'TTT':'TAG'] = result_wt_mm.loc[:, 'TTT':'TAG'].apply(lambda x: pd.to_numeric(x, errors='coerce'))

result_wt_mm['Gainers'] = (result_wt_mm.loc[:, result_wt_mm.columns.isin(gainers)].sum(axis=1)) / (result_wt_mm.loc[:, 'TTT':'TAG'].sum(axis=1))
result_wt_mm['Losers'] = (result_wt_mm.loc[:, result_wt_mm.columns.isin(losers)].sum(axis=1)) / (result_wt_mm.loc[:, 'TTT':'TAG'].sum(axis=1))
result_wt_mm

# процент каноничных генов
canon = result_wt_mm['GenName'].loc[result_wt_mm['GenName'].isin(['ND1', 'ND2', 'ND3', 'ND4', 'ND4L', 'ND5', 'ND6', 'COX1', 'COX2', 'COX3', 'CYTB', 'ATP6', 'ATP8'])].value_counts().sum()
print(canon * 100 / 130578)

result_wt_mm.to_csv('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/2Derived/08_CodonUsageDerived.csv', sep=';')

## Relative synonymous codon usage

rscu = pd.DataFrame(columns = ['Unnamed: 0', 'GenName', 'ProteinID', 'OrganismName', 'Taxonomy', 'Strand', 'TTT','TTC','TTA','TTG','TCT','TCC','TCA','TCG','TAT','TAC','TGT','TGC','TGA','TGG','CTT','CTC','CTA','CTG','CCT','CCC','CCA','CCG','CAT','CAC','CAA','CAG','CGT','CGC','CGA','CGG','ATT','ATC','ATA','ATG','ACT','ACC','ACA','ACG','AAT','AAC','AAA','AAG','AGT','AGC','AGA','AGG','GTT','GTC','GTA','GTG','GCT','GCC','GCA','GCG','GAT','GAC','GAA','GAG','GGT','GGC','GGA','GGG','TAA','TAG'])
direct2 = '/home/emulciber/MitoClub/mtDnaInvertebrate/Body/2Derived/codon_usage_rscu'
dirname2 = os.listdir(direct2)
for file in tqdm.tqdm(dirname2):
    try:
        df = pd.read_csv((os.path.join(direct2, file)), sep='\t')
        rscu = rscu.append(df, ignore_index=True)
    except:
        pass
rscu = rscu.loc[~rscu['ProteinID'].isin(exclude_list)]

rscu = rscu.merge(Tax, how='left', on='OrganismName')

rscu['Group3'] = 'Eukaryota'
rscu['Group3'].loc[rscu['Taxonomy'].str.contains('Fungi', na=False)] = 'Fungi'
rscu['Group3'].loc[rscu['Taxonomy'].str.contains('Viridiplantae', na=False)] = 'Plants'
rscu['Group3'].loc[rscu['Taxonomy'].str.contains('Metazoa', na=False)] = 'Invertebrates'
rscu['Group3'].loc[rscu['Taxonomy'].str.contains('Vertebrata', na=False)] = 'Vertebrata'
rscu['Group3'].loc[rscu['Taxonomy'].str.contains('Vertebrata' and 'Chondrichthyes', na=False)] = 'Chondrychthyes'
rscu['Group3'].loc[rscu['Taxonomy'].str.contains('Vertebrata' and 'Actinopterygii', na=False)] = 'Actinopterygii'
rscu['Group3'].loc[rscu['Taxonomy'].str.contains('Vertebrata' and 'Amphibia', na=False)] = 'Amphibia'
rscu['Group3'].loc[rscu['Taxonomy'].str.contains('Vertebrata' and 'Archelosauria')] = 'Reptilia'
rscu['Group3'].loc[rscu['Taxonomy'].str.contains('Vertebrata' and 'Lepidosauria', na=False)] = 'Reptilia'
rscu['Group3'].loc[rscu['Taxonomy'].str.contains('Vertebrata' and 'Aves', na=False)] = 'Aves'
rscu['Group3'].loc[rscu['Taxonomy'].str.contains('Vertebrata' and 'Mammalia', na=False)] = 'Mammalia'

rscu

rscu.to_csv('/home/emulciber/MitoClub/mtDnaInvertebrate/Body/2Derived/08_CodonUsageDerived_RSCU.csv', sep=';')
